// src/types/recycleBin/recycleBin.ts
export type RecycleBinType =
  | 'OrderRequest'
  | 'SalesOrder'
  | 'DeliveryNotes'
  | 'Quotations'
  | 'Sales'
  | 'SalesReturn'
  | 'Expenses'
  | 'PurchaseRequisition'
  | 'PurchaseOrder'
  | 'PurchasesReturn'
  | 'Purchases'
  | 'RecivingItems'
  | 'JobCards'
  | 'customer'
  | 'vendor'
  | 'employee';